Add tool type filter to list-images; rename package and module
