+++ .github/workflows/main.yml
@@ -25,8 +25,8 @@ jobs:
       - name: Test CLI
         run: |
           source /root/.bashrc
-          find_cli list-packages --test > packages.txt
+          aclimatise-automation list-packages --test > packages.txt
           mkdir out/
-          find_cli install packages.txt out/
-          find_cli wrappers out/
+          aclimatise-automation install packages.txt out/
+          aclimatise-automation wrappers out/
         shell: bash
